fix(autoFocus): Do not set focus outside of included content
Due to usage of `ng-include`, focus could be set to a control outside the partial just loaded.
Proceed to the end of current $digest before search for a control to focus.
If the directive is attached to an ui-view element, then listen to $viewContentLoaded events, otherwise listen to $includeContentLoaded events.

Fixes #2